+++ package.json
@@ -45,7 +45,7 @@
   },
   "homepage": "https://github.com/ws-node/astroboy.ts#readme",
   "dependencies": {
-    "@bonbons/di": "^1.4.1",
+    "@bonbons/di": "^1.4.1-rc.2",
     "@types/koa": "^2.0.47",
     "@types/koa-router": "^7.0.35",
     "@types/node": "^10.12.10",
