+++ src/utils.ts
@@ -3,6 +3,7 @@ import merge from "lodash/merge";
 import isPlainObject from "lodash/isPlainObject";
 import reduce from "lodash/reduce";
 import uuid from "uuid/v4";
+import chalk from "chalk";
 import {
   Constructor,
   getDependencies,
@@ -48,7 +49,15 @@ export function setColor(name: keyof typeof Colors, value: any): string {
   return `${Colors[name]}${value}${Colors.reset}`;
 }
 
-export const GlobalDI = new DIContainer<ScopeID, { ctx: any }>();
+export const GlobalDI = new DIContainer<ScopeID, { ctx: any }>({
+  throws: false,
+  onEmit: ({ level, data }) =>
+    console.log(
+      chalk[level === "warn" ? "yellow" : level === "error" ? "red" : "white"](
+        `[${level.toUpperCase()}] - ${data.msg}`
+      )
+    )
+});
 export const GlobalImplements = new Map<any, any>();
 
 export function setScopeId(ctx: IContext) {
